+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/plugin-bff
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/bff-core@2.48.3
+- @modern-js/create-request@2.48.3
+- @modern-js/server-utils@2.48.3
+- @modern-js/utils@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- Updated dependencies [ac586f1]
+  - @modern-js/create-request@2.48.2
+  - @modern-js/bff-core@2.48.2
+  - @modern-js/server-utils@2.48.2
+  - @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes