+++ packages/cli/plugin-swc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/core
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/utils@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes