+++ packages/generator/generators/monorepo-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/monorepo-generator
 
+## 3.3.14
+
+### Patch Changes
+
+- @modern-js/utils@2.48.3
+
+## 3.3.13
+
+### Patch Changes
+
+- @modern-js/utils@2.48.2
+
 ## 3.3.12
 
 ### Patch Changes