+++ packages/module/plugin-module-banner/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/plugin-module-banner
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/module-tools@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/module-tools@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes