+++ packages/server/babel-plugin-module-resolver/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/babel-plugin-module-resolver
 
+## 2.48.3
+
+## 2.48.2
+
 ## 2.48.1
 
 ## 2.48.0