+++ packages/server/create-request/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @modern-js/create-request
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.48.3
+- @modern-js/utils@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- ac586f1: fix(create-request): 修复错误的 types 指向
+  - @modern-js/runtime-utils@2.48.2
+  - @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes