+++ packages/solutions/app-tools/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @modern-js/app-tools
 
+## 2.48.3
+
+### Patch Changes
+
+- 18e6249: chore(server): set devServer.compress config value in app-tools
+
+  chore(server): 在 app-tools 中设置 ssr 场景下 devServer compress 配置的值
+
+- Updated dependencies [879f90b]
+- Updated dependencies [18e6249]
+  - @modern-js/uni-builder@2.48.3
+  - @modern-js/server@2.48.3
+  - @modern-js/core@2.48.3
+  - @modern-js/plugin-data-loader@2.48.3
+  - @modern-js/plugin-i18n@2.48.3
+  - @modern-js/plugin-lint@2.48.3
+  - @modern-js/server-core@2.48.3
+  - @modern-js/prod-server@2.48.3
+  - @modern-js/server-utils@2.48.3
+  - @modern-js/node-bundle-require@2.48.3
+  - @modern-js/plugin@2.48.3
+  - @modern-js/types@2.48.3
+  - @modern-js/utils@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- Updated dependencies [9e7115f]
+  - @modern-js/uni-builder@2.48.2
+  - @modern-js/server@2.48.2
+  - @modern-js/core@2.48.2
+  - @modern-js/plugin-data-loader@2.48.2
+  - @modern-js/plugin-i18n@2.48.2
+  - @modern-js/plugin-lint@2.48.2
+  - @modern-js/server-core@2.48.2
+  - @modern-js/prod-server@2.48.2
+  - @modern-js/server-utils@2.48.2
+  - @modern-js/node-bundle-require@2.48.2
+  - @modern-js/plugin@2.48.2
+  - @modern-js/types@2.48.2
+  - @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes