+++ packages/solutions/module-tools/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @modern-js/module-tools
 
+## 2.48.3
+
+### Patch Changes
+
+- @modern-js/core@2.48.3
+- @modern-js/plugin-changeset@2.48.3
+- @modern-js/plugin-i18n@2.48.3
+- @modern-js/plugin-lint@2.48.3
+- @modern-js/plugin@2.48.3
+- @modern-js/types@2.48.3
+- @modern-js/utils@2.48.3
+
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/core@2.48.2
+- @modern-js/plugin-changeset@2.48.2
+- @modern-js/plugin-i18n@2.48.2
+- @modern-js/plugin-lint@2.48.2
+- @modern-js/plugin@2.48.2
+- @modern-js/types@2.48.2
+- @modern-js/utils@2.48.2
+
 ## 2.48.1
 
 ### Patch Changes